+++ Game/Data/Troop/TroopTemplate.cs
@@ -103,8 +103,8 @@ namespace Game.Data.Troop {
                                        new DbColumn("type", battleStats.Base.Type, DbType.UInt16),
                                        new DbColumn("level", battleStats.Base.Lvl, DbType.Byte),
                                        new DbColumn("max_hp", battleStats.MaxHp, DbType.UInt16),
-                                       new DbColumn("attack", battleStats.Atk, DbType.Byte),
-                                       new DbColumn("defense", battleStats.Def, DbType.Byte),
+                                       new DbColumn("attack", battleStats.Atk, DbType.UInt16),
+                                       new DbColumn("defense", battleStats.Def, DbType.UInt16),
                                        new DbColumn("range", battleStats.Rng, DbType.Byte),
                                        new DbColumn("stealth", battleStats.Stl, DbType.Byte),
                                        new DbColumn("speed", battleStats.Spd, DbType.Byte)
